Fixes so default world is created if no specific one exists

+++ installServers.py
@@ -19,10 +19,11 @@ if len(sys.argv) == 1:
             game_version = data[serverName]["game_version"]
             install.installServer(serverName, game_version, str(25556 + i))
             
-            src = 'worlds/' + serverName
-        
-            dst = 'Servers/' + serverName + '/world'
-            shutil.copytree(src, dst, dirs_exist_ok=True)
+            if os.path.exists('/worlds/' + serverName):
+                src = 'worlds/' + serverName
+            
+                dst = 'Servers/' + serverName + '/world'
+                shutil.copytree(src, dst, dirs_exist_ok=True)
 else:
     servers = sys.argv
     with open('serverList.json', 'r') as file:
@@ -33,7 +34,8 @@ else:
                 game_version = data[serverName]["game_version"]
                 install.installServer(serverName, game_version, str(25556 + i))
                 
-                src = 'worlds/' + serverName
-            
-                dst = 'Servers/' + serverName + '/world'
-                shutil.copytree(src, dst, dirs_exist_ok=True)
+                if os.path.exists('/worlds/' + serverName):
+                    src = 'worlds/' + serverName
+                
+                    dst = 'Servers/' + serverName + '/world'
+                    shutil.copytree(src, dst, dirs_exist_ok=True)
